Refactored commands into separate class files into their own package
