+++ src/main/java/dk/xam/jbang/cli/JbangRun.java
@@ -1,0 +1,408 @@
+package dk.xam.jbang.cli;
+
+import static java.lang.System.*;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+import java.security.cert.X509Certificate;
+import java.util.*;
+import java.util.jar.Attributes;
+import java.util.jar.JarFile;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import javax.net.ssl.*;
+
+import org.apache.commons.text.StringEscapeUtils;
+
+import dk.xam.jbang.ExitException;
+import dk.xam.jbang.Script;
+import dk.xam.jbang.Settings;
+import dk.xam.jbang.Util;
+
+import picocli.CommandLine;
+
+@CommandLine.Command(name = "run", description = "Compile and run provided .java/.jsh script.")
+public class JbangRun extends JbangBaseScriptCommand {
+
+	@CommandLine.Option(names = { "-d",
+			"--debug" }, fallbackValue = "4004", parameterConsumer = IntFallbackConsumer.class, description = "Launch with java debug enabled on specified port (default: ${FALLBACK-VALUE}) ")
+	int debugPort = -1;
+
+	@CommandLine.Option(names = {
+			"--cds" }, description = "If specified Class Data Sharing (CDS) will be used for building and running (requires Java 13+)", negatable = true)
+	Boolean cds;
+
+	Optional<Boolean> cds() {
+		return Optional.ofNullable(cds);
+	}
+
+	boolean debug() {
+		return debugPort >= 0;
+	}
+
+	@CommandLine.Option(names = { "-D" }, description = "set a system property")
+	Map<String, String> properties = new HashMap<>();
+
+	@CommandLine.Option(names = {
+			"--insecure" }, description = "Enable insecure trust of all SSL certificates.", defaultValue = "false")
+	boolean insecure;
+
+	@CommandLine.Option(names = { "--interactive" }, description = "activate interactive mode")
+	boolean interactive;
+
+	@CommandLine.Option(names = {
+			"-n", "--native" }, description = "Build via native-image and run", defaultValue = "false")
+	boolean nativeImage;
+
+	@Override
+	public Integer doCall() throws IOException {
+		if (insecure) {
+			enableInsecure();
+		}
+
+		script = prepareScript(scriptOrFile);
+
+		if (script.needsJar()) {
+			build(script);
+		}
+		String cmdline = generateCommandLine(script);
+		if (isVerbose()) {
+			info("run: " + cmdline);
+		}
+		out.println(cmdline);
+
+		return 0;
+	}
+
+	private void enableInsecure() {
+		try {
+			// Create a trust manager that does not validate certificate chains
+			TrustManager[] trustAllCerts = new TrustManager[] { new X509TrustManager() {
+				public X509Certificate[] getAcceptedIssuers() {
+					return null;
+				}
+
+				public void checkClientTrusted(X509Certificate[] certs, String authType) {
+				}
+
+				public void checkServerTrusted(X509Certificate[] certs, String authType) {
+				}
+			}
+			};
+
+			// Install the all-trusting trust manager
+			SSLContext sc = SSLContext.getInstance("SSL");
+			sc.init(null, trustAllCerts, new java.security.SecureRandom());
+			HttpsURLConnection.setDefaultSSLSocketFactory(sc.getSocketFactory());
+
+			// Create all-trusting host name verifier
+			HostnameVerifier allHostsValid = new HostnameVerifier() {
+				public boolean verify(String hostname, SSLSession session) {
+					return true;
+				}
+			};
+
+			// Install the all-trusting host verifier
+			HttpsURLConnection.setDefaultHostnameVerifier(allHostsValid);
+		} catch (NoSuchAlgorithmException | KeyManagementException ex) {
+			throw new RuntimeException(ex);
+		}
+	}
+
+	// build with javac and then jar... todo: split up in more testable chunks
+	void build(Script script) throws IOException {
+		File baseDir = new File(Settings.getCacheDir().toFile(), "jars");
+		File tmpJarDir = new File(baseDir, script.backingFile.getName() +
+				"." + getStableID(script.backingFile));
+		tmpJarDir.mkdirs();
+
+		File outjar = new File(tmpJarDir.getParentFile(), tmpJarDir.getName() + ".jar");
+
+		if (!outjar.exists()) {
+			List<String> optionList = new ArrayList<String>();
+			optionList.add(resolveInJavaHome("javac"));
+			optionList.addAll(script.collectCompileOptions());
+			String path = script.resolveClassPath(offline);
+			if (!path.trim().isEmpty()) {
+				optionList.addAll(Arrays.asList("-classpath", path));
+			}
+			optionList.addAll(Arrays.asList("-d", tmpJarDir.getAbsolutePath()));
+			optionList.addAll(Arrays.asList(script.backingFile.getPath()));
+
+			Util.info("Building jar...");
+			if (isVerbose())
+				Util.info("compile: " + String.join(" ", optionList));
+
+			Process process = new ProcessBuilder(optionList).inheritIO().start();
+			try {
+				process.waitFor();
+			} catch (InterruptedException e) {
+				throw new ExitException(1, e);
+			}
+
+			if (process.exitValue() != 0) {
+				throw new ExitException(1, "Error during compile");
+			}
+
+			try {
+				// using Files.walk method with try-with-resources
+				try (Stream<Path> paths = Files.walk(tmpJarDir.toPath())) {
+					List<Path> items = paths.filter(Files::isRegularFile)
+											.filter(f -> !f.toFile().getName().contains("$"))
+											.filter(f -> f.toFile().getName().endsWith(".class"))
+											.collect(Collectors.toList());
+
+					if (items.size() > 1) { // todo: this feels like a very sketchy way to find the proper class name
+						// but it works.
+						String mainname = script.backingFile.getName().replace(".java", ".class");
+						items = items	.stream()
+										.filter(f -> f.toFile().getName().equalsIgnoreCase(mainname))
+										.collect(Collectors.toList());
+					}
+
+					if (items.size() != 1) {
+						throw new ExitException(1,
+								"Could not locate unique class. Found " + items.size() + " candidates.");
+					} else {
+						Path classfile = items.get(0);
+						String mainClass = findMainClass(tmpJarDir.toPath(), classfile);
+						script.setMainClass(mainClass);
+					}
+				}
+			} catch (IOException e) {
+				throw new ExitException(1, e);
+			}
+			script.createJarFile(tmpJarDir, outjar);
+
+		} else {
+			try (JarFile jf = new JarFile(outjar)) {
+				script.setMainClass(
+						jf.getManifest().getMainAttributes().getValue(Attributes.Name.MAIN_CLASS));
+			}
+		}
+
+		if (nativeImage && !getImageName(outjar).exists()) {
+			List<String> optionList = new ArrayList<String>();
+			optionList.add(resolveInGraalVMHome("native-image"));
+
+			optionList.add("-H:+ReportExceptionStackTraces");
+
+			optionList.add("--enable-https");
+
+			String classpath = script.resolveClassPath(offline);
+			optionList.add("--class-path=" + classpath);
+
+			optionList.add("-jar");
+			optionList.add(outjar.toString());
+
+			optionList.add(getImageName(outjar).toString());
+
+			File nilog = File.createTempFile("jbang", "native-image");
+			if (isVerbose())
+				Util.info("native-image: " + String.join(" ", optionList));
+			Util.info("log: " + nilog.toString());
+
+			Process process = new ProcessBuilder(optionList).inheritIO().redirectOutput(nilog).start();
+			try {
+				process.waitFor();
+			} catch (InterruptedException e) {
+				throw new ExitException(1, e);
+			}
+
+			if (process.exitValue() != 0) {
+				throw new ExitException(1, "Error during native-image");
+			}
+		}
+		script.setJar(outjar);
+	}
+
+	String generateCommandLine(Script script) throws IOException {
+
+		List<String> fullArgs = new ArrayList<>();
+
+		if (nativeImage) {
+			String imagename = getImageName(script.getJar()).toString();
+			if (new File(imagename).exists()) {
+				fullArgs.add(imagename);
+			} else {
+				warn("native built image not found - running in java mode.");
+			}
+		}
+
+		if (fullArgs.isEmpty()) {
+			String classpath = script.resolveClassPath(offline);
+
+			List<String> optionalArgs = new ArrayList<String>();
+
+			String javacmd = resolveInJavaHome("java");
+			if (script.backingFile.getName().endsWith(".jsh")) {
+
+				javacmd = resolveInJavaHome("jshell");
+				if (!classpath.trim().isEmpty()) {
+					optionalArgs.add("--class-path=" + classpath);
+				}
+
+				optionalArgs.add("--startup=DEFAULT");
+
+				File tempFile = File.createTempFile("jbang_arguments_", script.backingFile.getName());
+				Util.writeString(tempFile.toPath(), generateArgs(userParams, properties));
+
+				optionalArgs.add("--startup=" + tempFile.getAbsolutePath());
+
+				if (debug()) {
+					info("debug not possible when running via jshell.");
+				}
+
+			} else {
+				addPropertyFlags("-D", optionalArgs);
+
+				// optionalArgs.add("--source 11");
+				if (debug()) {
+					optionalArgs.add("-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=" + debugPort);
+				}
+				if (!classpath.trim().isEmpty()) {
+					optionalArgs.add("-classpath " + classpath);
+				}
+
+				if (optionActive(cds(), script.enableCDS())) {
+					String cdsJsa = script.getJar().getAbsolutePath() + ".jsa";
+					if (script.wasJarCreated()) {
+						if (isVerbose()) {
+							info("CDS: Archiving Classes At Exit at " + cdsJsa);
+						}
+						optionalArgs.add("-XX:ArchiveClassesAtExit=" + cdsJsa);
+					} else {
+						if (isVerbose()) {
+							info("CDS: Using shared archive classes from " + cdsJsa);
+						}
+						optionalArgs.add("-XX:SharedArchiveFile=" + cdsJsa);
+					}
+				}
+			}
+
+			// protect against spaces in dirs in paths, especially on windows.
+			if (javacmd.contains(" ")) {
+				javacmd = '"' + javacmd + '"';
+			}
+			fullArgs.add(javacmd);
+			fullArgs.addAll(script.collectRuntimeOptions());
+			fullArgs.addAll(script.getAutoDetectedModuleArguments(javacmd, offline));
+			fullArgs.addAll(optionalArgs);
+			if (script.getMainClass() != null) {
+				fullArgs.add(script.getMainClass());
+			} else {
+				fullArgs.add(script.backingFile.toString());
+			}
+		}
+
+		if (!script.forJShell()) {
+			fullArgs.addAll(userParams);
+		} else if (!interactive) {
+			File tempFile = File.createTempFile("jbang_exit_", script.backingFile.getName());
+			Util.writeString(tempFile.toPath(), "/exit");
+			fullArgs.add(tempFile.toString());
+		}
+
+		return String.join(" ", fullArgs);
+
+	}
+
+	static boolean optionActive(Optional<Boolean> master, boolean local) {
+		return master.map(Boolean::booleanValue).orElse(local);
+	}
+
+	private void addPropertyFlags(String def, List<String> optionalArgs) {
+		properties.forEach((k, e) -> {
+			optionalArgs.add(def + k + "=" + e);
+		});
+	}
+
+	/** based on jar what will the binary image name be. **/
+	private File getImageName(File outjar) {
+
+		if (getProperty("os.name").toLowerCase().startsWith("windows")) {
+			return new File(outjar.toString() + ".exe");
+		} else {
+			return new File(outjar.toString() + ".bin");
+		}
+	}
+
+	static public String findMainClass(Path base, Path classfile) {
+		StringBuilder mainClass = new StringBuilder(classfile.getFileName().toString().replace(".class", ""));
+		while (!classfile.getParent().equals(base)) {
+			classfile = classfile.getParent();
+			mainClass.insert(0, classfile.getFileName().toString() + ".");
+		}
+		return mainClass.toString();
+	}
+
+	String generateArgs(List<String> args, Map<String, String> properties) {
+
+		String buf = "String[] args = { " +
+				args.stream()
+					.map(s -> '"' + StringEscapeUtils.escapeJava(s) + '"')
+					.collect(Collectors.joining(", "))
+				+
+				" }" +
+				(properties.isEmpty() ? "" : "\n") +
+				properties	.entrySet()
+							.stream()
+							.map(x -> "System.setProperty(\"" + x.getKey() + "\",\"" + x.getValue() + "\");")
+							.collect(Collectors.joining("\n"));
+		return buf;
+	}
+
+	static String resolveInJavaHome(String cmd) {
+		return resolveInEnv("JAVA_HOME", cmd);
+	}
+
+	static String resolveInGraalVMHome(String cmd) {
+		String newcmd = resolveInEnv("GRAALVM_HOME", cmd);
+
+		if (newcmd.equals(cmd) &&
+				!new File(newcmd).exists()) {
+			return resolveInJavaHome(cmd);
+		} else {
+			return newcmd;
+		}
+	}
+
+	private static String resolveInEnv(String env, String cmd) {
+		if (getenv(env) != null) {
+			if (getProperty("os.name").toLowerCase().startsWith("windows")) {
+				cmd = cmd + ".exe";
+			}
+			return new File(getenv(env)).toPath().resolve("bin").resolve(cmd).toAbsolutePath().toString();
+		} else {
+			return cmd;
+		}
+	}
+
+	static class IntFallbackConsumer implements CommandLine.IParameterConsumer {
+		@Override
+		public void consumeParameters(Stack<String> args, CommandLine.Model.ArgSpec argSpec,
+				CommandLine.Model.CommandSpec commandSpec) {
+			String arg = args.pop();
+			try {
+				int port = Integer.parseInt(arg);
+				argSpec.setValue(port);
+			} catch (Exception ex) {
+				String fallbackValue = (argSpec.isOption()) ? ((CommandLine.Model.OptionSpec) argSpec).fallbackValue()
+						: null;
+				try {
+					int fallbackPort = Integer.parseInt(fallbackValue);
+					argSpec.setValue(fallbackPort);
+				} catch (Exception badFallbackValue) {
+					throw new CommandLine.InitializationException("FallbackValue for --debug must be an int",
+							badFallbackValue);
+				}
+				args.push(arg); // put it back
+			}
+		}
+	}
+}
