+++ src/test/java/dk/xam/jbang/TestUtil.java
@@ -1,0 +1,7 @@
+package dk.xam.jbang;
+
+public class TestUtil {
+	public static void clearSettingsAliasInfo() {
+		Settings.aliasInfo = null;
+	}
+}
